#1 rename files
for divide roles Japanese or English
