+++ src/english/ipa_phonetics.py
@@ -1,0 +1,9 @@
+from typing import List, Tuple
+
+
+class Phonetic:
+    """Add tags of HTML by pronunciations of English"""
+
+    def export_html(self, text_post: str) -> str:
+        pass
+
